+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### 1.1.1 - 12.01.2026
+- Fixed retention logic
+- Optimized for relative paths
+
 ### 1.1.0 - 11.01.2026
 - **BREAKING CHANGE:** If no files found by glob or regex a warning is printed instead of a raised error
 - Added `--fail-on-delete-error`